requirements.txt
openai
openai-agents
mcp
python-telegram-bot
python-dotenv
httpx
pydantic
requests